refactor(post_handler, post_store): change post ID type from int64 to int32 - update ID parsing in post_handler.go from 64-bit to 32-bit integer - convert ID to int32 when calling PostStore methods in post_handler.go - modify PostStore interface methods to use int32 for post IDs in post_store.go

+++ internal/store/post_store.go
@@ -8,10 +8,10 @@ import (
 
 type PostStore interface {
 	CreatePost(ctx context.Context, arg db.CreatePostParams) (db.Post, error)
-	GetPost(ctx context.Context, id int64) (db.Post, error)
+	GetPost(ctx context.Context, id int32) (db.Post, error)
 	ListPosts(ctx context.Context) ([]db.Post, error)
 	UpdatePost(ctx context.Context, arg db.UpdatePostParams) (db.Post, error)
-	DeletePost(ctx context.Context, id int64) error
+	DeletePost(ctx context.Context, id int32) error
 }
 
 type SQLPostStore struct {
